+++ actix-rt/CHANGES.md
@@ -1,7 +1,9 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Add `System::run_with_code` to allow retrieving the exit code on stop. [#411]
 
+[#411]: https://github.com/actix/actix-net/pull/411
 
 ## 2.4.0 - 2021-11-05
 * Add `Arbiter::try_current` for situations where thread may or may not have Arbiter context. [#408]